fix: organize shared packages
