+++ packages/data-model/src/index.ts
@@ -17,4 +17,5 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-export * from './types.js';
+export * from './study/study.js'; // Study
+export * from './study/studyTranslations.js'; // Study Translations
